pyproject.toml
[project]
name = "osago"
version = "0.1.0"
description = ""
authors = [
    {name = "ogyrec225",email = "vladkasimovskij34@gmail.com"}
]
packages = [{include = "osago", from = "src"}]
readme = "README.md"
requires-python = ">=3.11,<3.12"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
fastapi="^0.115.12"
pydantic="^2.11.4"
sqlalchemy="^2.0.40"
asyncpg="^0.30.0"
alembic="^1.15.2"
loguru="^0.7.3"
uvicorn="^0.34.2"
pydantic-settings = "^2.9.1"
dependency-injector = "^4.46.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.8"
pre-commit = "^4.2.0"
httpx = "^0.28.1"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.1.1"
pytest-dotenv = "^0.5.2"


[tool.ruff]
target-version = "py311"
line-length = 88
lint.select = [
    "E",      # Error
    "W",      # Warning
    "F",      # Pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "D",      # pydocstyle
    "UP",     # pyupgrade
]
lint.ignore = [
    "E501",   # line too long, handled by black
    "D100",   # Missing docstring in public module
    "D101",   # Missing docstring in public class
    "D102",   # Missing docstring in public method
    "D103",   # Missing docstring in public function
    "D104",   # Missing docstring in public package
    "D105",   # Missing docstring in magic method
    "D107",   # Missing docstring in __init__
    "D205",   # Missing blank line for summary 
    "B008",   # Ignore fastapi.(Depends...)
    "B904",   # Ignore nested raise
]
exclude = [".venv", "__pycache__", ".git"]